requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyJWT==2.8.0
azure-storage-blob==12.20.0
loguru==0.7.0
